scatteringmodel docs are good
