+++ docs/scatmodels.rst
@@ -3,54 +3,22 @@
 xdust.scatteringmodel
 =====================
 
-Abstract Class: *ScatteringModel*
----------------------------------
+The scattering model library is built on the superclass :ref:`ScatteringModelClass`, which defines the scattering physics model and holds the 
+computational results once the scattering model is calculated. 
 
-A dust scattering model must contain the following:
+**Premade scattering models**
 
-**Attributes:**
+* :ref:`RGscattering`
+* :ref:`Mie`
+* :ref:`GGADT`
+* :ref:`PAH`
 
-- ``stype`` -- a string describing the model
+.. _ScatteringModelClass:
 
-- ``cite`` -- a citation for the model--
+ScatteringModel
+----------------
 
-- ``qsca`` -- unitless scattering efficiency
-
-- ``qext`` -- unit-less extinction efficiency
-
-- ``qabs`` -- unit-less absorption efficiency
-
-- ``diff`` -- differential scattering efficiency, ster^-1
-
-- ``pars`` -- dict, stores the parameters used to run `calculate`
-
-**Methods**
-
-- ``calculate(lam, a, cm, mem_lim=8.0, theta=0.0)`` 
-    calculates the extinction 
-    efficiencies and differential scattering cross section for the given 
-    wavelength/energy grid, grain size, and composition.  
-    Stores the results in the attributes `qsca`, `qext`, `qabs`, and `diff`. 
-    The parameters used to run `calculate` are stored in the attribute `pars` 
-    as a dictionary.
-
-           ``lam`` -- astropy Quantity [wavelength or energy grid]
-
-           ``a``   -- float [grain size, micron] or astropy Quantity 
-
-           ``cm`` -- xdust.graindist.composition object
-
-           ``mem_lim`` -- float [memory limit for scattering calculation, GB, default 8.0]
-
-           ``theta`` -- float or np.array [radians] or astropy Quantity, angles to calculate differential scattering [default 0.0]
-
-- ``write_table(outfile)`` writes a FITS table of efficiency values to the specified filename   
-            
-            ``outfile`` -- string [filename for writing a FITS table of efficiency values]
-
-- ``read_from_table(infile)`` reads efficiency values from a FITS table and stores them in the attributes `qsca`, `qext`, `qabs`, and `diff`. The parameters used to run the scattering model are stored in the attribute `pars` as a dictionary. 
-
-            ``infile`` -- string [filename for loading efficiency values from FITS file]
+.. autoclass:: xdust.scatteringmodel.ScatteringModel
 
 
 .. _RGscattering:
@@ -70,3 +38,9 @@ Mie
 GGADT
 -----
 .. autoclass:: xdust.scatteringmodel.GGADT
+
+.. _PAH:
+
+PAH
+---
+.. autoclass:: xdust.scatteringmodel.PAH
