+++ src/xdust/scatteringmodel/rgscat.py
@@ -1,7 +1,7 @@
 import numpy as np
 import astropy.units as u
 from .. import helpers
-from .scatteringmodel import ScatteringModel
+from . import ScatteringModel
 
 __all__ = ['RGscattering']
 
@@ -9,8 +9,8 @@ CHARSIG       = 1.04 * u.arcmin # characteristic scattering angle [arcmin E(keV)
 
 class RGscattering(ScatteringModel):
     """
-    Rayleigh-Gans scattering model. *See* Mauche & Gorenstein (1986), ApJ 302, 371; 
-    Smith & Dwek (1998), ApJ, 503, 831
+    Rayleigh-Gans scattering model. *See* `Mauche & Gorenstein (1986), ApJ 302, 371 <https://ui.adsabs.harvard.edu/abs/1986ApJ...302..371M/abstract>`_ and
+    `Smith & Dwek (1998), ApJ, 503, 831 <https://ui.adsabs.harvard.edu/abs/1998ApJ...503..831S/abstract>`_
     """
     def __init__(self, **kwargs):
         ScatteringModel.__init__(self, **kwargs)
@@ -34,6 +34,10 @@ class RGscattering(ScatteringModel):
 
         theta : astropy.units.Quantity or numpy.ndarray or float
             Scattering angles; plain values are assumed to be in radians.
+
+        Returns
+        -------
+        None
 
         Updates the ``qsca``, ``qext``, ``qabs``, and ``diff`` attributes.
         """
@@ -101,7 +105,7 @@ class RGscattering(ScatteringModel):
         -------
         astropy.units.Quantity
             Characteristic scattering angle using
-            :math:`\\sigma = 1.04\\,\\mathrm{arcmin}\,(E/\\mathrm{keV})^{-1}(a/\\mu\\mathrm{m})^{-1}`.
+            :math:`\sigma = 1.04~\mathrm{arcmin}~(E/\mathrm{keV})^{-1}(a/\mu\mathrm{m})^{-1}`
         """
         lam_keV = lam
         if isinstance(lam, u.Quantity):
